Split apart serialization classes
