+++ algorithm/src/commonMain/kotlin/com/alexvanyo/composelife/model/FixedFormatCellStateSerializer.kt
@@ -1,0 +1,31 @@
+/*
+ * Copyright 2024 The Android Open Source Project
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.alexvanyo.composelife.model
+
+/**
+ * A serializer for a [CellStateFormat.FixedFormat].
+ *
+ * This can directly and synchronously deserialize a sequence of lines into a [DeserializationResult] with
+ * [deserializeToCellState], and serialize a [CellState] into a sequence of lines with [serializeToString].
+ */
+interface FixedFormatCellStateSerializer {
+    val format: CellStateFormat.FixedFormat
+
+    fun deserializeToCellState(lines: Sequence<String>): DeserializationResult
+
+    fun serializeToString(cellState: CellState): Sequence<String>
+}
